improving dashboard tests.
