+++ k2s/test/e2e/addons/dashboard/01-enabledisable/dashboard_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-package dashboard
+package dashboardenabdisable
 
 import (
 	"context"
