+++ k2s/test/e2e/addons/dashboard/02-securityenhanced/dashboard_security_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-package dashboardsecurity
+package dashboardsecurityenhanced
 
 import (
 	"context"
@@ -18,7 +18,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-const testClusterTimeout = time.Minute * 20
+const testClusterTimeout = time.Minute * 30
 
 var (
 	suite      *framework.K2sTestSuite
